leaderboard/apps/home/urls.py
from django.urls import path

from . import views

urlpatterns = [
    path('', views.Index.as_view()),
    # path('articles/<int:year>/', views.year_archive),
    # path('articles/<int:year>/<int:month>/', views.month_archive),
    # path('articles/<int:year>/<int:month>/<slug:slug>/', views.article_detail),
]