leaderboard/requirements/development.txt
-r base.txt

autopep8
pycodestyle
prospector
django-debug-toolbar
